requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
google-generativeai==0.3.2
openai>=1.3.7
qdrant-client==1.7.0
numpy==1.24.3
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1 

httpx>=0.27,<0.28
google-generativeai>=0.7,<0.8
Jinja2>=3.1,<3.2
pandas>=2.2,<3.0
